rename routes constant file
