+++ src/components/product-card/ProductCard.component.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 
-import { ROUTES } from '@/constants/Routes.ts';
+import { ROUTES } from '@/constants/Routes.constant.ts';
 import { CartContext } from '@/context/Cart.context.tsx';
 import { RouterContext } from '@/context/Router.context.tsx';
 import type { CartItem } from '@/interfaces/CartItem.interface.ts';
