+++ src/context/Router.context.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useEffect, useState } from 'react';
 
-import { ROUTES } from '../constants/Routes.ts';
+import { ROUTES } from '../constants/Routes.constant.ts';
 import { buildRoutePath, parseLocationHash, prepareRoutePathParameters } from '../helpers/RouterContext.helper.ts';
 import type { Route } from '../interfaces/Route.interface.ts';
 import type { RouterContextProps } from '../interfaces/RouterContextProps.interface.ts';
